// e-commerce_backoffice/src/models/settings/ShopInfos.ts
export interface ShopInfos{
    shop_name: string;
    shop_description: string;
    shop_logo_url: string;
    shop_email: string;
    shop_phone: string;
    shop_address_number: string;
    shop_address_street: string;
    shop_address_zipcode: string;
    shop_address_city: string;
    shop_address_country: string;
    shop_address_state: string;
    shop_language: string;
    shop_currency: string;
    shop_timezone: string;
    shop_weight_unit: string;
}